Making sure we use CLRs Math.Abs, no need to invent the wheel.

+++ src/SoundFingerprinting/Utils/FingerprintDescriptor.cs
@@ -57,7 +57,7 @@ namespace SoundFingerprinting.Utils
             for (int i = 0; i < topWavelets; i++)
             {
                 int index = indexes[i];
-                double value = concatenated[i];
+                float value = concatenated[i];
                 if (value > 0)
                 {
                     // positive wavelet
